# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PurchaseDetailOut": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "html_title": {
            "default": "",
            "title": "Html Title",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PurchaseDetailOut",
        "type": "object"
      },
      "PurchaseItemOut": {
        "properties": {
          "create_date": {
            "default": "",
            "title": "Create Date",
            "type": "string"
          },
          "creator_name": {
            "default": "",
            "title": "Creator Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_name": {
            "default": "",
            "title": "Project Name",
            "type": "string"
          },
          "project_no": {
            "default": "",
            "title": "Project No",
            "type": "string"
          },
          "requisition_no": {
            "default": "",
            "title": "Requisition No",
            "type": "string"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PurchaseItemOut",
        "type": "object"
      },
      "PurchaseListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurchaseItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 10,
            "title": "Page Size",
            "type": "integer"
          },
          "source_url": {
            "default": "",
            "title": "Source Url",
            "type": "string"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          },
          "total_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Text"
          },
          "view": {
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "view",
          "items"
        ],
        "title": "PurchaseListOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Browser-independent read-only facade for the authorized CTF target. Write operations are preview-only and never sent upstream. Local Bearer tokens never include upstream .iPSA cookie values.",
    "title": "iSStech Purchase Requisition Replay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/purchase-requisitions": {
      "get": {
        "operationId": "list_purchase_requisitions_v1_purchase_requisitions_get",
        "parameters": [
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "application",
              "title": "View",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project_no",
            "required": false,
            "schema": {
              "default": "",
              "title": "Project No",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requisition_no",
            "required": false,
            "schema": {
              "default": "",
              "title": "Requisition No",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_field",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort Field"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Purchase Requisitions",
        "tags": [
          "purchase-requisitions"
        ]
      }
    },
    "/v1/purchase-requisitions/{requisition_id}": {
      "get": {
        "operationId": "get_purchase_requisition_v1_purchase_requisitions__requisition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Purchase Requisition",
        "tags": [
          "purchase-requisitions"
        ]
      }
    },
...
  }
}